add light and dark buttons
